+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- `AdaptBasedOnVelocity`: try to reduce number of rejected timesteps.
+  For this, we now avoid increasing the *a priori* `dt` too abruptly.
+
 ## [0.20.0] - 2024-05-15
 
 ### Fixed