+++ CHANGELOG.md
@@ -5,6 +5,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.20.1] - 2024-05-16
+
 ### Changed
 
 - `AdaptBasedOnVelocity`: try to reduce number of rejected timesteps.